added popup into layout module
