+++ ClientAdvisor/App/frontend/src/pages/layout/Layout.module.css
@@ -149,7 +149,6 @@
   background: linear-gradient(180deg, #FFF 59.5%, #EBF3FC 100%);
   border-radius: 8px;
   height: 100%;
-  position: relative;
   box-shadow: 0px 0.83px 3.33px 0px #00000052;
   margin: 4rem 1rem 1rem;
 }
@@ -255,4 +254,58 @@
   padding: 10px 20px;
   cursor: pointer;
   font-size: 16px;
+}
+
+/* popup css*/
+.popupContainer {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  background: #fff;
+  border: 1px solid #ccc;
+  padding: 10px;
+  border-radius: 5px;
+  box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
+  z-index: 1;
+  position: fixed;
+  right: 1.4rem;
+  top: 4.4rem;
+}
+
+
+.popupContent {
+  display: flex;
+  align-items: center;
+}
+
+.checkmark {
+  color: white;
+  font-size: 13px;
+  vertical-align: middle;
+  padding-right: 0.4rem;
+}
+
+.popupText {
+  margin-left: 10px;
+}
+
+.headerText{
+  font-weight: 600;
+  color: #242424;
+}
+.popupSubtext {
+  font-size: smaller;
+  color: #242424;
+  padding-left: 1.4rem;
+  width: 75%;
+
+}
+
+.closeButton {
+  background: none;
+  border: none;
+  font-size: 13px;
+  cursor: pointer;
+  position: absolute;
+  right: 0.8rem;
 }
